requirements.txt
numpy>=2.0
opencv-python-headless>=4.13
Pillow>=12.0
python-pptx>=1.0
